+++ .flake8
@@ -2,3 +2,5 @@
 ignore = E121,E126,W503,W504
 exclude = mage_ai/data_preparation/templates,mage_ai/orchestration/db/migrations/versions
 max-line-length = 100
+per-file-ignores =
+    mage_ai/orchestration/pipeline_scheduler.py: F401,F403
